v1.1.0

+++ index.js
@@ -18,13 +18,33 @@ const lock     = new Lock(config.get('lockTimeout') * 1000, timestampFactory)
 const channels = initChannels(gpio, config.get('channels'), timestampFactory)
 const server   = mkServer({ name, logger, lock, channels })
 
+Object.keys(server.router.routes).map(key => {
+  server.router.routes[key].forEach(route => {
+    server.log.info({
+      method: route.spec.method,
+      path: route.spec.path
+    }, 'addRoute')
+  })
+})
+
 server.listen(port, () => {
   logger.info(`${server.name} listening on ${server.url}`)
+  logger.info('Now serving propane... and propane accessories')
 })
 
 setInterval(() => resetChannels(channels, timeout), 1000)
 
-process.on('exit', () => {
-  logger.info('Shutting down')
-  destroyChannels()
-})
+process.stdin.resume()
+process.on('SIGINT', exitHandler.bind(null, { exit: true }))
+process.on('uncaughtException', exitHandler.bind(null, { exit: true }))
+process.on('exit', exitHandler.bind(null, { cleanup: true }))
+
+function exitHandler({ clean, exit }, err) {
+  if (exit) {
+    if (err) logger.error(err, 'Shutting down')
+    else logger.warn('Shutting down')
+  }
+
+  if (clean) destroyChannels()
+  if (exit) process.exit(err ? 1 : 0)
+}
